tasks for week1

+++ week 1/task2/task2.c
@@ -1,9 +1,9 @@
 //Write a C code that will ask the user to enter a value then print it.
 #include<stdio.h>
 int main(){
-	int num;
+	float num;
 	printf("hello user enter any number: ");
-    scanf("%d",&num);	
-    printf("number is %d",num);
+    scanf("%g",&num);	
+    printf("number is %g",num);
 	return 0;
 }
